Add files via upload
In recent years, car accidents are the leading cause of death, killing around 1.3 million people
each year. Most of these accidents are caused by either driver distraction or drowsiness. This
number is still on the rise. So, it is very important to detect the drowsiness of the driver in order
to save the lives of the driver as well as the passengers.
This project mainly focuses on detecting the drowsiness and the inattentive level of the driver, to
decrease the car accident rate. The main goal is to provide an interface where the software can
automatically detect the drowsiness of the driver. This software is a real-time system that
captures images continuously and measures the state of the eye according to the specified
algorithm and gives a warning under necessary conditions.
The software captures the image of the person with the help of a webcam and examines with the
help of OpenCV which is an image processing tool and an open-source library which is used to
perform tasks like face detection, object tracking and many more. After the collection of live
images from the webcam stream the software then later applies the machine learning algorithm
using Python and Keras along with TensorFlow to the image to check whether the driver is
drowsy or not.
The detection of drowsiness is based on the per closure value of the eye and when the closure of
the eye exceeds a certain amount it senses that the driver is sleepy and plays the buzzer. Hence,
this project goes beyond the idea of detecting drowsiness while driving by notifying it.